+++ src/repl/views/terminal_renderer.rs
@@ -519,7 +519,7 @@ impl ViewRenderer for TerminalRenderer {
                     Show
                 )?;
             }
-            EditorMode::GMode => {
+            EditorMode::GPrefix => {
                 // Block cursor for G mode (same as normal mode)
                 tracing::debug!(
                     "render_cursor: showing cursor at ({}, {}) for G mode",
@@ -566,7 +566,7 @@ impl ViewRenderer for TerminalRenderer {
                 EditorMode::Normal => "NORMAL",
                 EditorMode::Insert => "INSERT",
                 EditorMode::Command => "COMMAND", // Shouldn't reach here
-                EditorMode::GMode => "NORMAL",    // Show as NORMAL since it's a micro mode
+                EditorMode::GPrefix => "NORMAL",  // Show as NORMAL since it's a prefix mode
             };
 
             let pane_text = match view_model.get_current_pane() {
@@ -647,7 +647,7 @@ impl ViewRenderer for TerminalRenderer {
             EditorMode::Normal => "NORMAL",
             EditorMode::Insert => "INSERT",
             EditorMode::Command => "COMMAND",
-            EditorMode::GMode => "NORMAL", // Show as NORMAL since it's a micro mode
+            EditorMode::GPrefix => "NORMAL", // Show as NORMAL since it's a prefix mode
         };
 
         let pane_text = match view_model.get_current_pane() {